+++ tests/test_all.nim
@@ -8,4 +8,5 @@ import t_core_types
 import t_threat_actor_models
 import t_config
 import t_parser
+import t_stream_engine
 import test1